pyproject.toml
[project]
name = "probing-jp-politeness"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.10,<3.12"
authors = [
    "Shusuke Hashimoto <shuhashi@iu.edu>",
    "Victor Shi <wencshi@iu.edu>"
]
dependencies = [
  "torch>2.5.1",
  "numpy<2",
  "pandas",
  "scikit-learn>=1.2,<1.5",
  "pyyaml",
  "requests",
  "tqdm",
  "transformers>=4.49,<5",
  "sentencepiece",
  "fugashi",
  "unidic_lite",
  "matplotlib>=3.10.8",
]
